# docs/swagger.yaml
basePath: /
definitions:
...
  internal_controllers.AdminOrderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_controllers.Order'
        type: array
      meta:
        $ref: '#/definitions/internal_controllers.MetaInfo'
    type: object
...
  internal_controllers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  internal_controllers.LoginInput:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  internal_controllers.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 3600
        type: integer
      user:
        $ref: '#/definitions/internal_controllers.UserResponse'
    type: object
  internal_controllers.MessageResponse:
    properties:
      message:
        example: success
        type: string
    type: object
  internal_controllers.MetaInfo:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  internal_controllers.Order:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      order_items:
        items:
          $ref: '#/definitions/internal_controllers.OrderItem'
        type: array
      status:
        example: pending
        type: string
      total_price:
        example: 1999.98
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  internal_controllers.OrderItem:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      order_id:
        example: 1
        type: integer
      price:
        example: 999.99
        type: number
      product:
        $ref: '#/definitions/internal_controllers.Product'
      product_id:
        example: 1
        type: integer
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 1999.98
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  internal_controllers.OrderStatsResponse:
    properties:
      counts:
        additionalProperties:
          format: int64
          type: integer
        example:
          confirmed: 20
          pending: 10
        type: object
      revenue:
        example: 50000
        type: number
    type: object
...
  internal_controllers.Product:
    properties:
      category:
        example: Electronics
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: High-performance laptop
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/image.jpg
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 10
        type: integer
      slug:
        example: laptop
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  internal_controllers.RefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 3600
        type: integer
    type: object
  internal_controllers.ResponseUser:
    properties:
      message:
        example: user created
        type: string
    type: object
  internal_controllers.SignupInput:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      role:
        example: customer
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  internal_controllers.UpdateOrderStatusInput:
    properties:
      status:
        example: confirmed
        type: string
    required:
    - status
    type: object
...
  internal_controllers.UserResponse:
    properties:
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      role:
        example: customer
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@ecommerce.com
    name: API Support
  description: This is a backend service for the Go e-commerce project with comprehensive
    authentication, product management, cart, orders, and payment functionality.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
  /api/admin/orders:
    get:
      description: Gets a paginated list of all orders with optional filters
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.AdminOrderListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all orders (Admin only)
      tags:
      - Admin
  /api/admin/orders/{id}:
    get:
      description: Retrieves detailed information about a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.Order'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID (Admin only)
      tags:
      - Admin
  /api/admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Updates the status of an order with validation
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateOrderStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update order status (Admin only)
      tags:
      - Admin
  /api/admin/orders/stats:
    get:
      description: Retrieves order counts by status and revenue summary
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.OrderStatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order statistics (Admin only)
      tags:
      - Admin
...
  /api/auth/logout:
    post:
      description: Clears refresh token and removes cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns access token with refresh cookie
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Rotates refresh token and issues new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.SignupInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_controllers.ResponseUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      summary: Register new user
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"